Create DAY1.cpp
Objective
In this challenge, we review some basic concepts that will get you started with this series. You will need to use the same (or similar) syntax to read input and write output in challenges throughout HackerRank. Check out the Tutorial tab for learning materials and an instructional video!

Task
To complete this challenge, you must save a line of input from stdin to a variable, print Hello, World. on a single line, and finally print the value of your variable on a second line.

Input Format

A single line of text denoting input_String  (the variable whose contents must be printed).

Output Format

Print Hello, World. on the first line, and the contents of input_String on the second line.